questions and results
Added the functionality to select the answers, only allowing one per question at a time.
added the functionality to show the results and highlight the correct answers in green and the selected incorrect answers in red.

need to style the results paragraph and also add option to the landing page to allow the player to select a category and difficulty level.

need to change check answer button after it has been click to give the option to play with the same settings or to return to the landing page in order to select new topics and difficulty.